requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scipy>=1.11.0
scikit-image>=0.21.0
tqdm>=4.66.0
pyyaml>=6.0
transformers>=4.35.0
huggingface-hub>=0.19.0
einops>=0.7.0
kornia>=0.7.0
albumentations>=1.3.0
simple-lama-inpainting>=0.1.0
